requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchain-openai
python-dotenv
aiofiles
openai
langchain-core